Plant_Disease_Detection/requirements.txt
tensorflow = 2.16.2
numpy
pandas
pillow
jupyter/colab
